requirements.txt
python-telegram-bot==21.11.1
python-dotenv==1.0.0
httpx==0.27
pydantic==2.10.6
loguru==0.7.2 